Make app interective, use modules
