+++ src/index.html
@@ -59,6 +59,6 @@
       <footer></footer>
     </main>
 
-    <script src="./script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
